+++ src/ska_pss_protest/candlist.py
@@ -167,6 +167,9 @@ class SpCcl:
         if len(cand_metadata) == 0:
             raise EOFError("Candidate list {} empty".format(cand_file))
         logging.info("Located {} candidates".format(len(cand_metadata)))
+
+        # Sort by S/N in descending order
+        cand_metadata.sort(key=lambda x: x[3], reverse=True)
         return cand_metadata
 
     @staticmethod