+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 4.0.4
+
+- All output data products (e.g., candidates, candidate metdata, log files, cheetah configs) are written to the same location to provide test diagnostics to developers/users
+- Added option to preserve data products/test results on disk after test runs. Default behaviour is that ProTest will clean up all created files.
+- Provided option to reduce SPS candidate filterbank to json strings decribing headers
+- Resolved major bugfix: SPS candidates are correctly sorted
+
 ## 4.0.3
 
 - Update pyproject.toml to include Pypi source
+